+++ lib/segment/src/index/field_index/field_index_base/field_index_read_impl.rs
@@ -8,26 +8,25 @@ use super::field_index_read::FieldIndexRead;
 use super::payload_field_index::PayloadFieldIndexRead;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::facet_index::{FacetIndex, FacetIndexEnum};
-use crate::index::field_index::full_text_index::text_index::PayloadMatchQueryType;
 use crate::index::field_index::numeric_index::{NumericFieldIndex, NumericFieldIndexRead};
 use crate::index::field_index::{CardinalityEstimation, PayloadBlockCondition};
 use crate::index::query_optimization::optimized_filter::ConditionCheckerFn;
 use crate::index::query_optimization::rescore_formula::value_retriever::VariableRetrieverFn;
 use crate::telemetry::PayloadIndexTelemetry;
-use crate::types::{FieldCondition, Match, MatchPhrase, MatchText, MatchTextAny, PayloadKeyType};
+use crate::types::{FieldCondition, PayloadKeyType};
 
 impl PayloadFieldIndexRead for FieldIndex {
     fn count_indexed_points(&self) -> usize {
         match self {
-            FieldIndex::IntIndex(idx) => idx.inner().count_indexed_points(),
-            FieldIndex::DatetimeIndex(idx) => idx.inner().count_indexed_points(),
+            FieldIndex::IntIndex(idx) => idx.count_indexed_points(),
+            FieldIndex::DatetimeIndex(idx) => idx.count_indexed_points(),
             FieldIndex::IntMapIndex(idx) => idx.count_indexed_points(),
             FieldIndex::KeywordIndex(idx) => idx.count_indexed_points(),
-            FieldIndex::FloatIndex(idx) => idx.inner().count_indexed_points(),
+            FieldIndex::FloatIndex(idx) => idx.count_indexed_points(),
             FieldIndex::GeoIndex(idx) => idx.count_indexed_points(),
             FieldIndex::BoolIndex(idx) => idx.count_indexed_points(),
             FieldIndex::FullTextIndex(idx) => idx.count_indexed_points(),
-            FieldIndex::UuidIndex(idx) => idx.inner().count_indexed_points(),
+            FieldIndex::UuidIndex(idx) => idx.count_indexed_points(),
             FieldIndex::UuidMapIndex(idx) => idx.count_indexed_points(),
             FieldIndex::NullIndex(idx) => idx.count_indexed_points(),
         }
@@ -39,15 +38,15 @@ impl PayloadFieldIndexRead for FieldIndex {
         hw_counter: &'a HardwareCounterCell,
     ) -> OperationResult<Option<Box<dyn Iterator<Item = PointOffsetType> + 'a>>> {
         match self {
-            FieldIndex::IntIndex(idx) => idx.inner().filter(condition, hw_counter),
-            FieldIndex::DatetimeIndex(idx) => idx.inner().filter(condition, hw_counter),
+            FieldIndex::IntIndex(idx) => idx.filter(condition, hw_counter),
+            FieldIndex::DatetimeIndex(idx) => idx.filter(condition, hw_counter),
             FieldIndex::IntMapIndex(idx) => idx.filter(condition, hw_counter),
             FieldIndex::KeywordIndex(idx) => idx.filter(condition, hw_counter),
-            FieldIndex::FloatIndex(idx) => idx.inner().filter(condition, hw_counter),
+            FieldIndex::FloatIndex(idx) => idx.filter(condition, hw_counter),
             FieldIndex::GeoIndex(idx) => idx.filter(condition, hw_counter),
             FieldIndex::BoolIndex(idx) => idx.filter(condition, hw_counter),
             FieldIndex::FullTextIndex(idx) => idx.filter(condition, hw_counter),
-            FieldIndex::UuidIndex(idx) => idx.inner().filter(condition, hw_counter),
+            FieldIndex::UuidIndex(idx) => idx.filter(condition, hw_counter),
             FieldIndex::UuidMapIndex(idx) => idx.filter(condition, hw_counter),
             FieldIndex::NullIndex(idx) => idx.filter(condition, hw_counter),
         }
@@ -59,17 +58,15 @@ impl PayloadFieldIndexRead for FieldIndex {
         hw_counter: &HardwareCounterCell,
     ) -> OperationResult<Option<CardinalityEstimation>> {
         match self {
-            FieldIndex::IntIndex(idx) => idx.inner().estimate_cardinality(condition, hw_counter),
-            FieldIndex::DatetimeIndex(idx) => {
-                idx.inner().estimate_cardinality(condition, hw_counter)
-            }
+            FieldIndex::IntIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
+            FieldIndex::DatetimeIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
             FieldIndex::IntMapIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
             FieldIndex::KeywordIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
-            FieldIndex::FloatIndex(idx) => idx.inner().estimate_cardinality(condition, hw_counter),
+            FieldIndex::FloatIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
             FieldIndex::GeoIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
             FieldIndex::BoolIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
             FieldIndex::FullTextIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
-            FieldIndex::UuidIndex(idx) => idx.inner().estimate_cardinality(condition, hw_counter),
+            FieldIndex::UuidIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
             FieldIndex::UuidMapIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
             FieldIndex::NullIndex(idx) => idx.estimate_cardinality(condition, hw_counter),
         }
@@ -82,15 +79,15 @@ impl PayloadFieldIndexRead for FieldIndex {
         f: &mut dyn FnMut(PayloadBlockCondition) -> OperationResult<()>,
     ) -> OperationResult<()> {
         match self {
-            FieldIndex::IntIndex(idx) => idx.inner().for_each_payload_block(threshold, key, f),
-            FieldIndex::DatetimeIndex(idx) => idx.inner().for_each_payload_block(threshold, key, f),
+            FieldIndex::IntIndex(idx) => idx.for_each_payload_block(threshold, key, f),
+            FieldIndex::DatetimeIndex(idx) => idx.for_each_payload_block(threshold, key, f),
             FieldIndex::IntMapIndex(idx) => idx.for_each_payload_block(threshold, key, f),
             FieldIndex::KeywordIndex(idx) => idx.for_each_payload_block(threshold, key, f),
-            FieldIndex::FloatIndex(idx) => idx.inner().for_each_payload_block(threshold, key, f),
+            FieldIndex::FloatIndex(idx) => idx.for_each_payload_block(threshold, key, f),
             FieldIndex::GeoIndex(idx) => idx.for_each_payload_block(threshold, key, f),
             FieldIndex::BoolIndex(idx) => idx.for_each_payload_block(threshold, key, f),
             FieldIndex::FullTextIndex(idx) => idx.for_each_payload_block(threshold, key, f),
-            FieldIndex::UuidIndex(idx) => idx.inner().for_each_payload_block(threshold, key, f),
+            FieldIndex::UuidIndex(idx) => idx.for_each_payload_block(threshold, key, f),
             FieldIndex::UuidMapIndex(idx) => idx.for_each_payload_block(threshold, key, f),
             FieldIndex::NullIndex(idx) => idx.for_each_payload_block(threshold, key, f),
         }
@@ -102,72 +99,65 @@ impl PayloadFieldIndexRead for FieldIndex {
         hw_acc: HwMeasurementAcc,
     ) -> Option<ConditionCheckerFn<'a>> {
         match self {
-            FieldIndex::IntIndex(idx) => idx.inner().condition_checker(condition, hw_acc),
-            FieldIndex::DatetimeIndex(idx) => idx.inner().condition_checker(condition, hw_acc),
+            FieldIndex::IntIndex(idx) => idx.condition_checker(condition, hw_acc),
+            FieldIndex::DatetimeIndex(idx) => idx.condition_checker(condition, hw_acc),
             FieldIndex::IntMapIndex(idx) => idx.condition_checker(condition, hw_acc),
             FieldIndex::KeywordIndex(idx) => idx.condition_checker(condition, hw_acc),
-            FieldIndex::FloatIndex(idx) => idx.inner().condition_checker(condition, hw_acc),
+            FieldIndex::FloatIndex(idx) => idx.condition_checker(condition, hw_acc),
             FieldIndex::GeoIndex(idx) => idx.condition_checker(condition, hw_acc),
             FieldIndex::BoolIndex(idx) => idx.condition_checker(condition, hw_acc),
             FieldIndex::FullTextIndex(idx) => idx.condition_checker(condition, hw_acc),
-            FieldIndex::UuidIndex(idx) => idx.inner().condition_checker(condition, hw_acc),
+            FieldIndex::UuidIndex(idx) => idx.condition_checker(condition, hw_acc),
             FieldIndex::UuidMapIndex(idx) => idx.condition_checker(condition, hw_acc),
             FieldIndex::NullIndex(idx) => idx.condition_checker(condition, hw_acc),
         }
     }
-}
 
-impl FieldIndexRead for FieldIndex {
-    /// Try to check condition for a payload given a field index.
-    /// Required because some index parameters may influence the condition checking logic.
-    /// For example, full text index may have different tokenizers.
-    ///
-    /// Returns `None` if there is no special logic for the given index
-    /// returns `Some(true)` if condition is satisfied
-    /// returns `Some(false)` if condition is not satisfied
     fn special_check_condition(
         &self,
         condition: &FieldCondition,
         payload_value: &Value,
         hw_counter: &HardwareCounterCell,
     ) -> OperationResult<Option<bool>> {
-        Ok(match self {
-            FieldIndex::IntIndex(_) => None,
-            FieldIndex::DatetimeIndex(_) => None,
-            FieldIndex::IntMapIndex(_) => None,
-            FieldIndex::KeywordIndex(_) => None,
-            FieldIndex::FloatIndex(_) => None,
-            FieldIndex::GeoIndex(_) => None,
-            FieldIndex::BoolIndex(_) => None,
-            FieldIndex::FullTextIndex(index) => match &condition.r#match {
-                Some(Match::Text(MatchText { text })) => Some(index.check_payload_match(
-                    payload_value,
-                    text,
-                    PayloadMatchQueryType::Text,
-                    hw_counter,
-                )?),
-                Some(Match::Phrase(MatchPhrase { phrase })) => Some(index.check_payload_match(
-                    payload_value,
-                    phrase,
-                    PayloadMatchQueryType::Phrase,
-                    hw_counter,
-                )?),
-                Some(Match::TextAny(MatchTextAny { text_any })) => {
-                    Some(index.check_payload_match(
-                        payload_value,
-                        text_any,
-                        PayloadMatchQueryType::TextAny,
-                        hw_counter,
-                    )?)
-                }
-                Some(Match::Value(_) | Match::Any(_) | Match::Except(_)) | None => None,
-            },
-            FieldIndex::UuidIndex(_) => None,
-            FieldIndex::UuidMapIndex(_) => None,
-            FieldIndex::NullIndex(_) => None,
-        })
+        match self {
+            FieldIndex::IntIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+            FieldIndex::DatetimeIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+            FieldIndex::IntMapIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+            FieldIndex::KeywordIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+            FieldIndex::FloatIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+            FieldIndex::GeoIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+            FieldIndex::BoolIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+            FieldIndex::FullTextIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+            FieldIndex::UuidIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+            FieldIndex::UuidMapIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+            FieldIndex::NullIndex(idx) => {
+                idx.special_check_condition(condition, payload_value, hw_counter)
+            }
+        }
     }
+}
 
+impl FieldIndexRead for FieldIndex {
     fn get_telemetry_data(&self) -> PayloadIndexTelemetry {
         match self {
             FieldIndex::IntIndex(index) => index.get_telemetry_data(),
