+++ lib/segment/src/index/field_index/full_text_index/tests/mod.rs
@@ -8,7 +8,7 @@ use tempfile::Builder;
 use crate::data_types::index::{TextIndexParams, TextIndexType, TokenizerType};
 use crate::index::field_index::full_text_index::text_index::FullTextIndex;
 use crate::index::field_index::{
-    FieldIndex, FieldIndexBuilderTrait as _, FieldIndexRead, ValueIndexer,
+    FieldIndex, FieldIndexBuilderTrait as _, PayloadFieldIndexRead, ValueIndexer,
 };
 
 fn movie_titles() -> Vec<String> {
