+++ lib/segment/src/index/field_index/numeric_index/tests.rs
@@ -1,12 +1,26 @@
+use std::path::Path;
+
+use common::bitvec::{BitSlice, BitVec};
 use common::counter::hardware_accumulator::HwMeasurementAcc;
+use common::counter::hardware_counter::HardwareCounterCell;
+use common::types::PointOffsetType;
+use gridstore::Blob;
 use itertools::Itertools;
+use ordered_float::OrderedFloat;
 use rand::prelude::StdRng;
 use rand::{RngExt, SeedableRng};
 use rstest::rstest;
+use serde_json::Value;
 use tempfile::{Builder, TempDir};
 
+use super::immutable_numeric_index::ImmutableNumericIndex;
 use super::*;
+use crate::index::field_index::stored_point_to_values::StoredValue;
+use crate::index::field_index::{
+    CardinalityEstimation, FieldIndexBuilderTrait, PayloadFieldIndexRead, ValueIndexer,
+};
 use crate::json_path::JsonPath;
+use crate::types::FieldCondition;
 
 /// Generous default size for the deleted-points bitslice used in tests.
 ///
